Improve feature layout for TlsBraid
TlsBraid does not depend on feature = “stream”
