+++ src/stream/core.rs
@@ -2,9 +2,6 @@
 
 use std::pin::pin;
 
-#[cfg(feature = "tls")]
-use std::task::Poll;
-
 use pin_project::pin_project;
 use tokio::io::{AsyncRead, AsyncWrite};
 use tokio::net::{TcpStream, UnixStream};
@@ -13,9 +10,6 @@ use crate::stream::duplex::DuplexStream;
 use crate::stream::info::{ConnectionInfo, HasConnectionInfo};
 
 use super::info::BraidAddr;
-
-#[cfg(feature = "tls")]
-use super::tls::TlsHandshakeStream;
 
 /// Dispatching wrapper for potential stream connection types
 ///
@@ -110,96 +104,5 @@ impl From<DuplexStream> for Braid {
 impl From<UnixStream> for Braid {
     fn from(stream: UnixStream) -> Self {
         Self::Unix(stream)
-    }
-}
-
-/// Dispatching wrapper for optionally supporting TLS
-#[cfg(feature = "tls")]
-#[derive(Debug)]
-#[pin_project(project=BraidProjection)]
-pub enum TlsBraid<Tls, NoTls> {
-    /// A stream without TLS
-    NoTls(#[pin] NoTls),
-
-    /// A stream with TLS
-    Tls(#[pin] Tls),
-}
-
-#[cfg(feature = "tls")]
-impl<Tls, NoTls> TlsHandshakeStream for TlsBraid<Tls, NoTls>
-where
-    Tls: TlsHandshakeStream + Unpin,
-    NoTls: AsyncRead + AsyncWrite + Unpin,
-{
-    fn poll_handshake(
-        &mut self,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Result<(), std::io::Error>> {
-        match self {
-            TlsBraid::NoTls(_) => Poll::Ready(Ok(())),
-            TlsBraid::Tls(ref mut stream) => stream.poll_handshake(cx),
-        }
-    }
-}
-
-#[cfg(feature = "tls")]
-macro_rules! dispatch {
-    ($driver:ident.$method:ident($($args:expr),+)) => {
-
-        match $driver.project() {
-            BraidProjection::NoTls(stream) => stream.$method($($args),+),
-            BraidProjection::Tls(stream) => stream.$method($($args),+),
-        }
-    };
-}
-
-#[cfg(feature = "tls")]
-impl<Tls, NoTls> AsyncRead for TlsBraid<Tls, NoTls>
-where
-    Tls: AsyncRead,
-    NoTls: AsyncRead,
-{
-    fn poll_read(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> std::task::Poll<std::io::Result<()>> {
-        dispatch!(self.poll_read(cx, buf))
-    }
-}
-
-#[cfg(feature = "tls")]
-impl<Tls, NoTls> AsyncWrite for TlsBraid<Tls, NoTls>
-where
-    Tls: AsyncWrite,
-    NoTls: AsyncWrite,
-{
-    fn poll_write(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &[u8],
-    ) -> std::task::Poll<Result<usize, std::io::Error>> {
-        dispatch!(self.poll_write(cx, buf))
-    }
-
-    fn poll_flush(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Result<(), std::io::Error>> {
-        dispatch!(self.poll_flush(cx))
-    }
-
-    fn poll_shutdown(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Result<(), std::io::Error>> {
-        dispatch!(self.poll_shutdown(cx))
-    }
-}
-
-#[cfg(feature = "tls")]
-impl<Tls, NoTls> From<NoTls> for TlsBraid<Tls, NoTls> {
-    fn from(stream: NoTls) -> Self {
-        Self::NoTls(stream)
     }
 }
