+++ src/stream/tls/mod.rs
@@ -11,10 +11,96 @@ use std::{
 };
 
 pub use info::TlsConnectionInfo;
+use pin_project::pin_project;
 use tokio::io::{AsyncRead, AsyncWrite};
 
 /// A stream that supports a TLS handshake.
 pub trait TlsHandshakeStream: AsyncRead + AsyncWrite {
     /// Poll the handshake to completion.
     fn poll_handshake(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), io::Error>>;
+}
+
+/// Dispatching wrapper for optionally supporting TLS
+#[derive(Debug)]
+#[pin_project(project=BraidProjection)]
+pub enum TlsBraid<Tls, NoTls> {
+    /// A stream without TLS
+    NoTls(#[pin] NoTls),
+
+    /// A stream with TLS
+    Tls(#[pin] Tls),
+}
+
+impl<Tls, NoTls> TlsHandshakeStream for TlsBraid<Tls, NoTls>
+where
+    Tls: TlsHandshakeStream + Unpin,
+    NoTls: AsyncRead + AsyncWrite + Unpin,
+{
+    fn poll_handshake(
+        &mut self,
+        cx: &mut std::task::Context<'_>,
+    ) -> std::task::Poll<Result<(), std::io::Error>> {
+        match self {
+            TlsBraid::NoTls(_) => Poll::Ready(Ok(())),
+            TlsBraid::Tls(ref mut stream) => stream.poll_handshake(cx),
+        }
+    }
+}
+
+macro_rules! dispatch {
+    ($driver:ident.$method:ident($($args:expr),+)) => {
+
+        match $driver.project() {
+            BraidProjection::NoTls(stream) => stream.$method($($args),+),
+            BraidProjection::Tls(stream) => stream.$method($($args),+),
+        }
+    };
+}
+
+impl<Tls, NoTls> AsyncRead for TlsBraid<Tls, NoTls>
+where
+    Tls: AsyncRead,
+    NoTls: AsyncRead,
+{
+    fn poll_read(
+        self: std::pin::Pin<&mut Self>,
+        cx: &mut std::task::Context<'_>,
+        buf: &mut tokio::io::ReadBuf<'_>,
+    ) -> std::task::Poll<std::io::Result<()>> {
+        dispatch!(self.poll_read(cx, buf))
+    }
+}
+
+impl<Tls, NoTls> AsyncWrite for TlsBraid<Tls, NoTls>
+where
+    Tls: AsyncWrite,
+    NoTls: AsyncWrite,
+{
+    fn poll_write(
+        self: std::pin::Pin<&mut Self>,
+        cx: &mut std::task::Context<'_>,
+        buf: &[u8],
+    ) -> std::task::Poll<Result<usize, std::io::Error>> {
+        dispatch!(self.poll_write(cx, buf))
+    }
+
+    fn poll_flush(
+        self: std::pin::Pin<&mut Self>,
+        cx: &mut std::task::Context<'_>,
+    ) -> std::task::Poll<Result<(), std::io::Error>> {
+        dispatch!(self.poll_flush(cx))
+    }
+
+    fn poll_shutdown(
+        self: std::pin::Pin<&mut Self>,
+        cx: &mut std::task::Context<'_>,
+    ) -> std::task::Poll<Result<(), std::io::Error>> {
+        dispatch!(self.poll_shutdown(cx))
+    }
+}
+
+impl<Tls, NoTls> From<NoTls> for TlsBraid<Tls, NoTls> {
+    fn from(stream: NoTls) -> Self {
+        Self::NoTls(stream)
+    }
 }
